refactor codebase, create utilities folder
