+++ my-app/src/components/MenuIcon.tsx
@@ -2,6 +2,7 @@ import { useState } from "react";
 import React, { JSX } from "react";
 
 interface MenuIconProps {
+  onToggle: (newState: boolean) => void;
   // Add any props here if needed in the future
 }
 
@@ -9,7 +10,11 @@ export default function MenuIcon(props: MenuIconProps): JSX.Element {
   const [state, setState] = useState<boolean>(false);
 
   function toggleIcon(): void {
-    setState((prev: boolean) => !prev);
+    setState((prev: boolean) => {
+      const newState = !prev;
+      props.onToggle(newState);
+      return newState;
+    });
   }
 
   return (
@@ -24,5 +29,3 @@ export default function MenuIcon(props: MenuIconProps): JSX.Element {
     </>
   );
 }
-
-
